+++ packages/core/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+- Fix: add all objects in `resolver.of` to `context.types`
+
 ## 0.4.0 (2024-11-120)
 
 - Refactor: follow standard-schema ([#7](https://github.com/modevol-com/gqloom/pull/7))