+++ packages/core/CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
-- Fix: add all objects in `resolver.of` to `context.types`
+* Fix: add all objects in `resolver.of` to `context.types`
+* Chore: update `@standard-schema/spec` to 1.0.0-beta.4
 
 ## 0.4.0 (2024-11-120)
 
-- Refactor: follow standard-schema ([#7](https://github.com/modevol-com/gqloom/pull/7))
+* Refactor: follow standard-schema ([#7](https://github.com/modevol-com/gqloom/pull/7))
 
 ## 0.3.0 (2024-10-19)