+++ packages/core/CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+## 0.5.0 (2024-12-03)
+
 * Fix: add all objects in `resolver.of` to `context.types`
 * Chore: update `@standard-schema/spec` to 1.0.0-beta.4
 * Refactor: update `CallableMiddlewareOptions` for middleware
 
-## 0.4.0 (2024-11-120)
+## 0.4.0 (2024-11-12)
 
 * Refactor: follow standard-schema ([#7](https://github.com/modevol-com/gqloom/pull/7))